Added Under Development PopUp
Added a new view and methods to the PopUp Folder: Under Development.

The popUp is missing an image to show to the user. This can be assigned
directly in the .xib file of the popUps.

The popUp is shown on the main menu, when you press the “Mi progreso”
button.